Update front-end and cors fixed at myapiapp

+++ python/myapiapp.py
@@ -1,10 +1,12 @@
 from flask import Flask, request, jsonify
+from flask_cors import CORS
 import zipfile
 import os
 import filecutting
 import solution
 
 app = Flask (__name__)
+CORS(app)
 UPLOAD_FOLDER = 'plagiarism-checker/python/upload_zip/'
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 
@@ -20,16 +22,23 @@ def uploader():
         file.save(os.path.join(app.config['UPLOAD_FOLDER'], file.filename))
     
     items = os.listdir(UPLOAD_FOLDER)
-    with zipfile.ZipFile(UPLOAD_FOLDER + items[0], "r") as zip_ref:
-        zip_ref.extractall("plagiarism-checker/docs/words/")
-    
-    for x in items:
-        os.remove(UPLOAD_FOLDER + x)
-    
-    file_status = filecutting.runfilecutting()
-    records = solution.processing()
 
-    return jsonify(status = file_status, record = records)
+    if len(items) != None:
+        filecutting.runfilecutting()
+        records = solution.processing()
+
+        return jsonify(record = records)
+    else:
+        with zipfile.ZipFile(UPLOAD_FOLDER + items[0], "r") as zip_ref:
+            zip_ref.extractall("plagiarism-checker/docs/words/")
+        
+        for x in items:
+            os.remove(UPLOAD_FOLDER + x)
+        filecutting.runfilecutting()
+        records = solution.processing()
+
+        return jsonify(record = {records})
+    
 
 
 if __name__ == '__main__':
